When receiving a MIDI keypress, wait for a timeout before flushing LEDs, after clearing a LED light.
This is useful with multiple notes on a chord, as we get multiple subsequent MIDI key press events, but we can't flush per-event. So, after the first press of a chord, we wait for gathering possible further keypresses, before flushing. Here we're adding a FLUSH_TOUT latency from when a key is pressed and when its light is cleared (5ms at the moment).